refactor: routes & rename components
